Reapply "Merge branch 'transitionECS' into Dev4"
This reverts commit 28262e2eef47b812baeaa2734dfb8ccad244a1b9.

+++ Project-X/Tool/Animation.h
@@ -13,7 +13,7 @@ private:
     
     sf::Vector2i sizeSpriteSheet;
     int nbrFrames;
-    int frameRate;
+    float frameRate;
     
     sf::Vector2i locTexture;
     int actualFrame;
